add remove messages.

+++ deployment/node/sdu/sdu-opt-glassfish3/pre-remove.sh
@@ -1,4 +1,12 @@
 #!/bin/sh
 
-rm -fr /opt/java/glassfish-3.0.1
-rm -f  /opt/java/glassfish
+instdir=/opt/java/glassfish-3.0.1
+
+if [ -d $instdir ]; then
+
+    echo "Remove $instdir"
+    rm -fr /opt/java/glassfish-3.0.1
+
+    rm -f  /opt/java/glassfish
+
+fi
